player-windows/requirements.txt
PyQt6==6.7.1
PyQt6-WebEngine==6.7.0
python-mpv==1.0.7
requests==2.32.3
websocket-client==1.8.0
pywin32==306; sys_platform == "win32"
